+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## 7.0.0-alpha.3
+## Unreleased
 
 ### Features
 
 - Error Cause Extractor ([#1198](https://github.com/getsentry/sentry-dart/pull/1198))
+  - Add `throwable` to `SentryException`
+
+## 7.0.0-alpha.3
+
 ### Breaking Changes
 
 - Rename APM tracking feature flags to tracing ([#1222](https://github.com/getsentry/sentry-dart/pull/1222))